Refactor the StructuredDocument API: Split into interfaces and classes which are also moved into the ast.impl package
