+++ asciidoctorj-core/src/main/java/org/asciidoctor/ast/ContentPart.java
@@ -1,76 +1,25 @@
 package org.asciidoctor.ast;
 
+import java.util.*;
 import java.util.List;
-import java.util.Map;
 
-public class ContentPart {
-    private String id;
-    private int level;
-    private String context;
-    private String style;
-    private String role;
-    private String title;
-    private Map<String, Object> attributes;
-    private String content;
-    private List<ContentPart> parts;
+public interface ContentPart {
 
-    private ContentPart() {
-        super();
-    }
+    String getId();
 
-    public String getId() {
-        return id;
-    }
+    int getLevel();
 
-    public int getLevel() {
-        return level;
-    }
+    String getContext();
 
-    public String getContext() {
-        return context;
-    }
+    String getStyle();
 
-    public String getStyle() {
-        return style;
-    }
+    String getRole();
 
-    public String getRole() {
-        return role;
-    }
+    String getTitle();
 
-    public String getTitle() {
-        return title;
-    }
+    Map<String, Object> getAttributes();
 
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
+    String getContent();
 
-    public String getContent() {
-        return content;
-    }
-
-    public List<ContentPart> getParts() {
-        return parts;
-    }
-
-    public void setParts(List<ContentPart> parts) {
-        this.parts = parts;
-    }
-
-    public static ContentPart createContentPart(String id, int level, String context, String title, String style,
-            String role, Map<String, Object> attributes, String textContent) {
-        ContentPart contentPart = new ContentPart();
-        contentPart.level = level;
-        contentPart.id = id;
-        contentPart.context = context;
-        contentPart.style = style;
-        contentPart.role = role;
-        contentPart.title = title;
-        contentPart.attributes = attributes;
-        contentPart.content = textContent;
-
-        return contentPart;
-    }
-
+    List<? extends ContentPart> getParts();
 }
