+++ asciidoctorj-core/src/main/java/org/asciidoctor/ast/DocumentHeader.java
@@ -1,75 +1,18 @@
 package org.asciidoctor.ast;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
-import org.asciidoctor.internal.CaseInsensitiveMap;
+public interface DocumentHeader {
 
-public class DocumentHeader {
+    java.util.List<? extends Author> getAuthors();
 
-    private Title documentTitle;
-    private String pageTitle;
-    private Author author;
-    private List<Author> authors = new ArrayList<Author>();
-    private RevisionInfo revisionInfo;
+    Title getDocumentTitle();
 
-    private Map<String, Object> attributes;
+    String getPageTitle();
 
-    private DocumentHeader() {
-        super();
-    }
+    Author getAuthor();
 
-    public List<Author> getAuthors() {
-        return this.authors;
-    }
+    RevisionInfo getRevisionInfo();
 
-    public Title getDocumentTitle() {
-        return documentTitle;
-    }
-
-    public String getPageTitle() {
-        return pageTitle;
-    }
-
-    public Author getAuthor() {
-        return author;
-    }
-
-    public RevisionInfo getRevisionInfo() {
-        return revisionInfo;
-    }
-
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    public static DocumentHeader createDocumentHeader(Title documentTitle, String pageTitle,
-            Map<String, Object> attributes) {
-
-        DocumentHeader documentHeader = new DocumentHeader();
-
-        documentHeader.documentTitle = documentTitle;
-        documentHeader.pageTitle = pageTitle;
-        documentHeader.attributes = new CaseInsensitiveMap<String, Object>(attributes);
-
-        documentHeader.author = getAuthor(attributes);
-        documentHeader.revisionInfo = geRevisionInfo(attributes);
-        documentHeader.authors.addAll(getAuthors(attributes));
-
-        return documentHeader;
-    }
-
-    private static List<Author> getAuthors(Map<String, Object> attributes) {
-        return Author.getAuthors(attributes);
-    }
-
-    private static Author getAuthor(Map<String, Object> attributes) {
-        return Author.getInstance(attributes);
-    }
-
-    private static RevisionInfo geRevisionInfo(Map<String, Object> attributes) {
-        return RevisionInfo.getInstance(attributes);
-    }
-
+    Map<String, Object> getAttributes();
 }
