+++ asciidoctorj-core/src/main/java/org/asciidoctor/ast/RevisionInfo.java
@@ -1,58 +1,12 @@
 package org.asciidoctor.ast;
 
-import java.util.Map;
+/**
+ * Created by robertpanzer on 23.05.15.
+ */
+public interface RevisionInfo {
+    String getDate();
 
-public class RevisionInfo {
+    String getNumber();
 
-    private static String REV_DATE_ATTRIBUTE_NAME = "revdate";
-    private static String REV_NUMBER_ATTRIBUTE_NAME = "revnumber";
-    private static String REV_REMARK_ATTRIBUTE_NAME = "revremark";
-
-    private String date;
-    private String number;
-    private String remark;
-
-    public static RevisionInfo getInstance(Map<String, Object> attributes) {
-
-        RevisionInfo revisionInfo = new RevisionInfo();
-
-        if (attributes.containsKey(REV_DATE_ATTRIBUTE_NAME)) {
-            revisionInfo.setDate((String) attributes.get(REV_DATE_ATTRIBUTE_NAME));
-        }
-
-        if (attributes.containsKey(REV_NUMBER_ATTRIBUTE_NAME)) {
-            revisionInfo.setNumber((String) attributes.get(REV_NUMBER_ATTRIBUTE_NAME));
-        }
-
-        if (attributes.containsKey(REV_REMARK_ATTRIBUTE_NAME)) {
-            revisionInfo.setRemark((String) attributes.get(REV_REMARK_ATTRIBUTE_NAME));
-        }
-
-        return revisionInfo;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public String getNumber() {
-        return number;
-    }
-
-    public void setNumber(String number) {
-        this.number = number;
-    }
-
-    public String getRemark() {
-        return remark;
-    }
-
-    public void setRemark(String remark) {
-        this.remark = remark;
-    }
-
+    String getRemark();
 }
