+++ asciidoctorj-core/src/test/java/org/asciidoctor/WhenStructuredDocumentIsRequired.java
@@ -39,7 +39,7 @@ public class WhenStructuredDocumentIsRequired {
 		StructuredDocument structuredDocument = asciidoctor
 				.readDocumentStructure(s, new java.util.HashMap<String, Object>());
 		
-		List<ContentPart> parts = structuredDocument.getParts();
+		List<? extends ContentPart> parts = structuredDocument.getParts();
 		assertThat(parts, hasSize(1));
 
 		DocumentHeader header = structuredDocument.getHeader();
@@ -74,7 +74,7 @@ public class WhenStructuredDocumentIsRequired {
 		assertThat((String) attributes.get("email"),
 				is("doc.writer@asciidoc.org"));
 
-		List<ContentPart> parts = document.getParts();
+		List<? extends ContentPart> parts = document.getParts();
 
 		assertThat(parts, notNullValue());
 		assertThat(parts, hasSize(3));
@@ -120,7 +120,7 @@ public class WhenStructuredDocumentIsRequired {
 		Map<String, Object> attributes = header.getAttributes();
 		assertThat((String) attributes.get("type"), is("object.type"));
 
-		List<ContentPart> parts = document.getParts();
+		List<? extends ContentPart> parts = document.getParts();
 
 		assertThat(parts, notNullValue());
 		assertThat(parts, hasSize(4));
@@ -128,7 +128,7 @@ public class WhenStructuredDocumentIsRequired {
 		assertThat(document.getPartByStyle("literal").getContent(),
 				is("TODO: This is description"));
 
-		List<ContentPart> images = document.getPartsByContext("image");
+		List<? extends ContentPart> images = document.getPartsByContext("image");
 		assertThat(images, hasSize(2));
 		for (int i = 0; i < 2; i++) {
 			assertThat((String) images.get(i).getAttributes().get("target"),
@@ -152,7 +152,7 @@ public class WhenStructuredDocumentIsRequired {
 
 		assertThat(header.getDocumentTitle().getMain(), is("Sample Document"));
 
-		List<ContentPart> parts = document.getParts();
+		List<? extends ContentPart> parts = document.getParts();
 
 		assertThat(parts, notNullValue());
 		assertThat(parts, hasSize(0));
@@ -169,7 +169,7 @@ public class WhenStructuredDocumentIsRequired {
 
 		assertThat(header, notNullValue());
 
-		List<ContentPart> parts = document.getParts();
+		List<? extends ContentPart> parts = document.getParts();
 
 		assertThat(parts, notNullValue());
 		assertThat(parts, hasSize(1));
@@ -187,7 +187,7 @@ public class WhenStructuredDocumentIsRequired {
 
 		assertThat(header, notNullValue());
 
-		List<ContentPart> parts = document.getParts();
+		List<? extends ContentPart> parts = document.getParts();
 
 		assertThat(parts, notNullValue());
 		assertThat(parts, hasSize(0));
